+++ app/Database/Migrations/2021-05-01-001900_Posts.php
@@ -40,7 +40,7 @@ class Posts extends Migration
 			]
 		]);
 		$this->forge->addKey('id', true);
-		$this->forge->createTable('posts');
+		$this->forge->createTable('posts',true);
 	}
 
 	public function down()