+++ app/Views/admin/template.php
@@ -28,10 +28,7 @@
 						<a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
 					</li>
 					<li class="nav-item d-none d-sm-inline-block">
-						<a href="../../index3.html" class="nav-link">Home</a>
-					</li>
-					<li class="nav-item d-none d-sm-inline-block">
-						<a href="#" class="nav-link">Contact</a>
+						<a href="<?= base_url('/admin/') ?>" class="nav-link">Home</a>
 					</li>
 				</ul>
 				<!-- Right navbar links -->
@@ -158,8 +155,8 @@
 			<!-- Main Sidebar Container -->
 			<aside class="main-sidebar sidebar-dark-primary elevation-4">
 				<!-- Brand Logo -->
-				<a href="../../index3.html" class="brand-link">
-				<img src="../../adminlte/img/AdminLTELogo.png" alt="AdminLTE Logo" class="brand-image img-circle elevation-3" style="opacity: .8">
+				<a href="<?= base_url('/admin/') ?>" class="brand-link">
+				<img src="<?= base_url('/adminlte/img/AdminLTELogo.png') ?>" alt="AdminLTE Logo" class="brand-image img-circle elevation-3" style="opacity: .8">
 				<span class="brand-text font-weight-light">AdminLTE 3</span>
 				</a>
 				<!-- Sidebar -->
@@ -167,7 +164,7 @@
 					<!-- Sidebar user (optional) -->
 					<div class="user-panel mt-3 pb-3 mb-3 d-flex">
 						<div class="image">
-							<img src="../../adminlte/img/user2-160x160.jpg" class="img-circle elevation-2" alt="User Image">
+							<img src="<?= base_url('/adminlte/img/user2-160x160.jpg') ?>" class="img-circle elevation-2" alt="User Image">
 						</div>
 						<div class="info">
 							<a href="#" class="d-block">Alexander Pierce</a>
@@ -223,6 +220,27 @@
 									</li>
 								</ul>
 							</li>
+							<li class="nav-item">
+								<a href="#" class="nav-link">
+									<i class="nav-icon fas fa-user"></i>
+									<p>
+										Usuários
+										<i class="right fas fa-angle-left"></i>
+									</p>
+								</a>
+								<ul class="nav nav-treeview">
+									<li class="nav-item">
+										<a href="<?= base_url('/admin/users') ?>" class="nav-link">
+											<p>Todos os usuários</p>
+										</a>
+									</li>
+									<li class="nav-item">
+										<a href="<?= base_url('/admin/users/create') ?>" class="nav-link">
+											<p>Adicionar novo</p>
+										</a>
+									</li>
+								</ul>
+							</li>
 						</ul>
 					</nav>
 					<!-- /.sidebar-menu -->
@@ -256,7 +274,7 @@
 							  echo view($view);
 							}
 							?>
-					</div>
+					</div>					
 				</section>
 				<!-- /.content -->
 			</div>